+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.1.6 (2024-09-09)
 1. add filename_filter to utils.format_error
+2. add `functools.get_function` to find function from entrypoint
 
 ### 1.1.5 (2024-08-29)
 1. add `utils.get_hash_int`